Drop Xcode project, rename file extensions instead.
In order to gradually convert the code to using generic source inputs, I rename the extensions of Swift files to disable them from building and focus on one file at a time. I do this instead of using an Xcode project, as Xcode just keeps crashing and giving errors that do not occur when opening this project as a swift package. Both Xcode 11.6 and Xcode 12 beta.